pyproject.toml
[project]
name = "ocr-guia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "pyinstaller>=6.19.0",
    "pymupdf>=1.27.1",
    "pystray>=0.19.5",
    "rapidocr-onnxruntime>=1.4.4",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
    "dulwich"
]
